refactor: resolve circular dependencies and enhance build portability
- Fix circular imports in Beaufort, Autokey, and Running Key ciphers
- Implement adaptive n-gram scorer selection for better short-text support
- Replace sed with portable Node script in CJS build pipeline
- Strengthen Running Key test assertions
- Align Beaufort module structure with other ciphers
