+++ package.json
@@ -8,7 +8,7 @@
   "main": "./dist/cjs/index.js",
   "scripts": {
     "test": "cross-env TSX_TSCONFIG_PATH='./tsconfig.test.json' mocha",
-    "build:cjs": "tsc --project tsconfig.commonjs.json && tsc-alias -p tsconfig.commonjs.json && sed -i 's|import\\.meta\\.url|\"file:\/\/\"+__filename|g' dist/cjs/utils/cryptanalysis.js && node -e \"require('node:fs').cpSync('src/ngrams', 'dist/cjs/ngrams', {recursive:true})\"",
+    "build:cjs": "tsc --project tsconfig.commonjs.json && tsc-alias -p tsconfig.commonjs.json && node -e \"const fs = require('node:fs'); const path = 'dist/cjs/utils/cryptanalysis.js'; let content = fs.readFileSync(path, 'utf8'); content = content.replace(/import\\.meta\\.url/g, '\\\"file://\\\"+__filename'); fs.writeFileSync(path, content);\" && node -e \"require('node:fs').cpSync('src/ngrams', 'dist/cjs/ngrams', {recursive:true})\"",
     "build:esm": "tsc --project tsconfig.esm.json && tsc-alias -p tsconfig.esm.json && node -e \"require('node:fs').cpSync('src/ngrams', 'dist/esm/ngrams', {recursive:true})\"",
     "build:types": "tsc -p tsconfig.types.json",
     "build": "npm run build:cjs && npm run build:esm",
